add functions to draw graph without deleted sections

+++ main_program/autoDeleteGraph.py
@@ -1,23 +1,27 @@
 import os
 from typing import List
-
-from graph_Manager import Graph
+import re
+from graph_Manager import *
 from start_module.Variables import *
 from StaticMethods import *
 from Time import Time
 
 
-class deleteLabel:
-    def __init__(self, start_period: Time, finish_period: Time):
-        self.start_period = start_period
-        self.finish_period = finish_period
+class deleteAndFindLabel:
+    def __init__(self, start_period: Time, finish_period: Time, patient: str = None):
+        self.__start_period = start_period
+        self.__finish_period = finish_period
+        self.__patient = patient
 
-    def deletedPeriod(self):
-        return f"""DELETED FROM Time1:{str(self.start_period)} TO Time2:{str(self.finish_period)} PERIOD;\n"""
+    def genFileName(self):
+        return f"""/{self.__patient}#{str(self.__start_period)}#{str(self.__finish_period)}#deleted{FilesConstant.text_type}"""
+
+    def deletedPeriodLabel(self):
+        return f"""DELETED FROM Time1:{str(self.__start_period)} TO Time2:{str(self.__finish_period)} PERIOD;\n"""
 
 
 class Dir:
-    def __init__(self, patient, start_period, finish_period):
+    def __init__(self, patient, start_period: int, finish_period: int):
         self.patient = patient
         self.start_period = start_period
         self.finish_period = finish_period
@@ -26,7 +30,7 @@ class Dir:
         return f"""DELETED FROM T1:{sec_to_time_short(self.start_period)} TO T2:{sec_to_time_short(self.finish_period)} PERIOD;"""
 
     def __str__(self):
-        return f"""/{self.patient}_{sec_to_time_short(self.start_period)}@{sec_to_time_short(self.finish_period)}_deleted{FilesConstant.text_type}"""
+        return f"""/{self.patient}#{sec_to_time_short(self.start_period)}#{sec_to_time_short(self.finish_period)}#deleted{FilesConstant.text_type}"""
 
 
 class autoDeleteGraph:
@@ -38,7 +42,7 @@ class autoDeleteGraph:
 
         self.__graph_constant: Graph = Graph()
         self.__graph_constant.file_init()
-        self.__graph_constant.start_init()
+
         self.__first_init = False
         self.__need_to_delete = []
 
@@ -59,10 +63,27 @@ class autoDeleteGraph:
         v: str = self.__graph_constant.file_n
         return v.split("/")[-1]
 
-    def readData(self, file_dir: str):
-        print(f"""[INFO] trying to read file: {self.__file_direction} without data""")
+    def readData(self, _start: Time, _finish: Time):
+        current_path: str = self.__generateFileName(_start, _finish)
+        if not os.path.exists(current_path):
+            print(f"[INFO] file does not exist")
+            return None
+        x_s = []
+        y_s = []
+        for i in range(self.__graph_constant.getSignalsAmount()):
+            y_s.append([])
+        with open(current_path, 'r') as f:
+            for line in f:
+                if not line[0] == 'D':
+                    current_line = re.sub(r"\s{2,}", ";", line).replace("\n", "").split(";")
+                    if len(current_line) > 1:
+                        x_s.append(Time.parseOne(current_line[0]).get_seconds())
+                        for j in range(self.__graph_constant.getSignalsAmount()):
+                            y_s[j].append(float(current_line[j + 1]))
+        return [x_s, y_s]
 
     def writeData(self, start_time: Time = None, finish_time: Time = None):
+
 
         if start_time is None:
             self.__startTime = self.__graph_constant.current_xs[0]
@@ -73,7 +94,11 @@ class autoDeleteGraph:
             self.__finishTime = self.__graph_constant.current_xs[-1]
         else:
             self.__finishTime = finish_time.getSeconds()
-
+        s = self.__generateFileName(start_time, finish_time)
+        if os.path.exists(s):
+            raise Exception(
+                f"Forbidden to modify file. It is already exist; Delete file from folder or choose another input time;")
+        self.__graph_constant.start_init()
         self.__full_file_name = self.__file_direction + str(
             Dir(self.getPatientName(), self.__startTime, self.__finishTime))
         with open(self.__full_file_name, 'w') as current_file:
@@ -87,28 +112,34 @@ class autoDeleteGraph:
                     current_file.write(self.__generateOneLine(i))
                     i = i + 1
                     st = self.__graph_constant.getElementPos(j[1].getSeconds()) + 1
-                current_file.write("\n" + deleteLabel(j[0], j[1]).deletedPeriod() + "\n")
+                current_file.write("\n" + deleteAndFindLabel(j[0], j[1]).deletedPeriodLabel() + "\n")
             fin = self.__graph_constant.getElementPos(self.__finishTime)
             i = st
             while i <= fin:
                 current_file.write(self.__generateOneLine(i))
                 i = i + 1
-
+        print(f"[INFO] file:{self.__generateFileName(start_time, finish_time)} saved;")
         return True
 
     def __generateOneLine(self, i: int):
         current_line = f"{sec_to_time_short(self.__graph_constant.current_xs[i]):<30}"
         for j in range(self.__graph_constant.getSignalsAmount()):
-            current_line = current_line + f"{self.__graph_constant.getCurrentYS(j)[i]:<30}"
+            current_line = current_line + f"{self.__graph_constant.getCurrentYS(j)[i]:>30}"
         current_line = current_line + "\n"
         return current_line
+
+    def __generateFileName(self, startTime: Time, finishTime: Time):
+        d = deleteAndFindLabel(startTime, finishTime, self.getPatientName()).genFileName()
+        return self.__file_direction + "/" + str(d)
 
     def deleteData(self, what_to_delete: List = []):
         self.__need_to_delete = what_to_delete
         return True
 
-    def make_graph(self, input_file_name: str):
-        pass
+    def make_graph(self, start_time: Time, finish_time: Time):
+        xs1, ys1 = self.readData(start_time, finish_time)
+        self.__graph_constant.start_init(xs1, ys1)
+        start_plot(self.__graph_constant)
 
     def get_current_file_dir(self):
         return self.__file_direction
@@ -117,9 +148,10 @@ class autoDeleteGraph:
 if __name__ == "__main__":
     autoDelGraph = autoDeleteGraph()
     start = Time()
-    need_to_del = []
-    need_to_del.append([Time(0, 0, 2, 0), Time(0, 0, 8, 0)])
-    need_to_del.append([Time(0, 0, 9, 0), Time(0, 0, 10, 0)])
-    autoDelGraph.deleteData(need_to_del)
-    finish = Time(0, 0, 10, 0)
-    autoDelGraph.writeData(start, finish)
+    # need_to_del = []
+    # need_to_del.append([Time(0, 0, 2, 0), Time(0, 0, 8, 0)])
+    # need_to_del.append([Time(0, 0, 9, 0), Time(1, 0, 0, 0)])
+    # autoDelGraph.deleteData(need_to_del)
+    finish = Time(2, 0, 0, 0)
+    # autoDelGraph.writeData(start, finish)
+    autoDelGraph.make_graph(start, finish)
